requirements.txt
altgraph==0.17
APScheduler==3.6.3
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
auto-py-to-exe==2.7.8
beautifulsoup4==4.10.0
bitstring==3.1.9
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
cffi==1.14.3
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==3.3.1
cycler==0.10.0
Cython==0.29.26
decorator==4.4.2
docopt==0.6.2
Eel==0.12.4
eth-hash==0.3.2
eth-typing==2.3.0
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
isort==5.6.4
Js2Py==0.71
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.2
packaging==20.4
pefile==2019.4.18
Pillow==8.0.0
pipwin==0.5.1
pluggy==0.13.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.9.0
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyjsparser==2.7.1
pylibscrypt==2.0.0
pylint==2.6.0
pyodbc==4.0.32
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pytest==6.1.1
python-dateutil==2.8.1
python-telegram-bot==12.7
pytz==2020.5
pywin32==228
pywin32-ctypes==0.2.0
requests==2.27.1
rfc3986==1.5.0
scrypt==0.8.19
six==1.15.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
tk==0.1.0
tk-helper==1.0.1
tkinter-fonts-viewer==0.1.1
toml==0.10.1
toolz==0.11.2
tornado==6.1
ttkwidgets==0.12.1
tzlocal==2.1
ujson==5.1.0
urllib3==1.26.8
voluptuous==0.12.2
w3lib==1.22.0
wcwidth==0.2.5
whichcraft==0.6.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.1.2
